refactor: InputView, OutputView 패키지 변경 -> views
